style: Move MediaList to its own module.
And move the parsing from a free function to MediaList::parse.

Bug: 1468846
MozReview-Commit-ID: 75ES6I2EEOE
